Revert "Attempt to ignore templator tests for Windows"
This reverts commit ca71069685b69aca6b926590d73bf738814852d3.
